[#1835] Reorganize AbilityData & SkillData, convert SpellData to plain schema
